backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-ollama>=0.3.6
faiss-cpu>=1.7.4
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
python-multipart>=0.0.6
aiofiles>=23.0.0 